Project: Divide project into two bins and a lib
Refactor project source files into two binaries that can run the
project, a gui and a cli, and an underlying library that they utilize.
This is done to support code reusability, extensibility and
separation of concerns.

Note that this commit breaks the gui. The gui had a known issue in which
it hard-coded cli options in order to run and could not be run from
cargo. This commit will make it easier for the gui to directly use the
library code to run counters instead.